pi/requirements.txt
bme280pi==1.1.0
paho_mqtt==1.5.1
